Move Value class to Orange.data.variable module
