+++ Orange/data/__init__.py
@@ -1,5 +1,4 @@
 from .instance import *
-from .value import *
 from .variable import *
 from .domain import *
 from .storage import *
